SegundaParte
Creo que termine con Security. Faltaria inicializar la bdd con los usuarios, permisos y roles para poder usar la api.

+++ src/main/java/ExperienceGroup/Ludora/features/client/domain/dto/ClientDTOResponse.java
@@ -7,8 +7,8 @@ public record ClientDTOResponse (UUID externalId,
                                  String name,
                                  String lastName,
                                  String userName,
-                                 int phone,
+                                 Long phone,
                                  String street,
-                                 int numberStreet,
+                                 Integer numberStreet,
                                  LocalDate birthDate)
 { }
